ぜ

+++ MisakiAobaBot.py
@@ -112,7 +112,10 @@ def start(bot, update):
 def help(bot, update):
     """Send a message when the command /help is issued."""
     if update.message.date > init_time:
-        bot.send_message(chat_id=update.message.chat_id, text=GLOBAL_WORDS.word_help, 
+        if randrange(1000)<30:
+            bot.send_message(chat_id=update.message.chat_id, text="ぜ")
+        else:
+            bot.send_message(chat_id=update.message.chat_id, text=GLOBAL_WORDS.word_help, 
                         parse_mode=ParseMode.HTML)
 
 def tbgame(bot, update):
@@ -125,10 +128,13 @@ def tbgame(bot, update):
 def rule(bot, update):
     """Send a message when the command /rule is issued."""
     if update.message.date > init_time:
-        msg=bot.send_message(chat_id=update.message.chat_id, text=GLOBAL_WORDS.word_rule, 
-                        parse_mode=ParseMode.HTML)
-        time.sleep(60)
-        bot.delete_message(chat_id=update.message.chat_id, message_id=msg.message_id)
+        if randrange(1000)<30:
+            bot.send_message(chat_id=update.message.chat_id, text="ぜ")
+        else:
+            msg=bot.send_message(chat_id=update.message.chat_id, text=GLOBAL_WORDS.word_rule, 
+                            parse_mode=ParseMode.HTML)
+            time.sleep(60)
+            bot.delete_message(chat_id=update.message.chat_id, message_id=msg.message_id)
 
 def state(bot, update):
     """Send a message when the command /state is issued."""
